rename parsimonious -> persistence
