+++ scripts/experiments/03_best_performance.py
@@ -3,7 +3,7 @@ import sys
 sys.path.append("../..")
 
 from scripts.utils import _rename_directory, get_data_path
-from _base_models import parsimonious, regression, linear_nn, rnn, earnn
+from _base_models import persistence, regression, linear_nn, rnn, earnn
 
 
 if __name__ == "__main__":
@@ -13,7 +13,7 @@ if __name__ == "__main__":
     important_vars = ["VCI", "precip", "t2m", "pev", "E", "SMsurf", "SMroot"]
     always_ignore_vars = ["ndvi", "p84.162", "sp", "tp", "Eb"]
 
-    # parsimonious()
+    # persistence()
     # regression(ignore_vars=always_ignore_vars)
     # gbdt(ignore_vars=always_ignore_vars)
     # linear_nn(ignore_vars=always_ignore_vars)
