+++ scripts/experiments/07_nowcast.py
@@ -4,7 +4,7 @@ sys.path.append("../..")
 
 from scripts.utils import _rename_directory, get_data_path
 from srcipts.engineer import engineer
-from _base_models import parsimonious, regression, linear_nn, rnn, earnn
+from _base_models import persistence, regression, linear_nn, rnn, earnn
 
 
 def run_engineer() -> None:
@@ -33,7 +33,7 @@ if __name__ == "__main__":
         "tprate_mean_3",
     ]
 
-    parsimonious(experiment="nowcast")
+    persistence(experiment="nowcast")
     # regression(ignore_vars=always_ignore_vars)
     # gbdt(ignore_vars=always_ignore_vars)
     # linear_nn(ignore_vars=always_ignore_vars)
