+++ scripts/experiments/_base_models.py
@@ -14,7 +14,7 @@ from src.analysis import all_explanations_for_file
 from scripts.utils import get_data_path
 
 
-def parsimonious(experiment="one_month_forecast",):
+def persistence(experiment="one_month_forecast",):
     predictor = Persistence(get_data_path(), experiment=experiment)
     predictor.evaluate(save_preds=True)
