+++ scripts/models.py
@@ -16,7 +16,7 @@ from src.analysis import all_explanations_for_file
 from scripts.utils import get_data_path
 
 
-def parsimonious(experiment="one_month_forecast",):
+def persistence(experiment="one_month_forecast",):
 
     data_path = get_data_path()
     predictor = Persistence(
@@ -126,7 +126,7 @@ def earnn(
 
 if __name__ == "__main__":
     ignore_vars = ["VCI", "p84.162", "sp", "tp"]
-    parsimonious(ignore_vars=ignore_vars)
+    persistence(ignore_vars=ignore_vars)
     regression(ignore_vars=ignore_vars)
     linear_nn(ignore_vars=ignore_vars)
     rnn(ignore_vars=ignore_vars)
